# Improve performance of column cloning inside DataFrame arithmetics