--- a/src/Microsoft.Data.Analysis/ReadOnlyDataFrameBuffer.cs
+++ b/src/Microsoft.Data.Analysis/ReadOnlyDataFrameBuffer.cs
@@ -48,15 +48,16 @@ namespace Microsoft.Data.Analysis
             get => (MemoryMarshal.Cast<byte, T>(ReadOnlyBuffer.Span)).Slice(0, Length);
         }
 
-        public int Length { get; internal set; }
+        public int Length { get; protected set; }
 
-        public ReadOnlyDataFrameBuffer(int numberOfValues = 8)
+        public ReadOnlyDataFrameBuffer(int length = 0)
         {
-            if ((long)numberOfValues * Size > MaxCapacity)
+            if ((long)length * Size > MaxCapacity)
             {
-                throw new ArgumentException($"{numberOfValues} exceeds buffer capacity", nameof(numberOfValues));
+                throw new ArgumentException($"{length} exceeds buffer capacity", nameof(length));
             }
-            _readOnlyBuffer = new byte[numberOfValues * Size];
+            _readOnlyBuffer = new byte[length * Size];
+            Length = length;
         }
 
         public ReadOnlyDataFrameBuffer(ReadOnlyMemory<byte> buffer, int length)