The goal of this PR is to perform Arithmetics operation on columns with the same underlying data type approximately 3 times faster.

Detail of changes:

1) Fix PrimitiveColumnContainer Clone() method to use memory block coping for internal buffer instead of appending values one by one (with memory reallocation on each buffer resizing cycle). Do similar changes for CloneNullBitMapBuffers() method

2)  Improve BinaryOperation.Implementation methods for all Arithmetic operations that happen not in place (default behavior). 
Before the change autogenerated code looked like this:

```
public partial class SingleDataFrameColumn
{
    internal SingleDataFrameColumn AddImplementation(SingleDataFrameColumn column, bool inPlace = false)
    {
        if (column.Length != Length)
        {
            throw new ArgumentException(Strings.MismatchedColumnLengths, nameof(column));
        }
        SingleDataFrameColumn newColumn = inPlace ? this : CloneAsSingleColumn();
        newColumn.ColumnContainer.Add(column.ColumnContainer);
        return newColumn;
    }
}
```

After PR https://github.com/dotnet/machinelearning/pull/6677 CloneAsSingleColumn can be changed to just this.Clone(). This allow to avoid unnecessary type conversion, that happens inside CloneAs... method and use fast Clone() method with bulk memory copy for internal buffers. For example. for Single:

```
internal PrimitiveColumnContainer<float> CloneAsFloatContainer()
{
 ...
    for (int i = 0; i < span.Length; i++)
   {
          newBuffer.Append(SingleConverter<T>.Instance.GetSingle(span[i]);
   }
}
```

3) Fix DataFrameBuffer constructor. 
DataFrameBuffer overrides parent ReadOnlyDataFrameBuffer ReadOnlyBuffer to return own new field _memory instead of parent _readOnlyBuffer (after this parent _readonlybuffer is ignored and never used). However in constructor _memory is not created, instead base constructor is called to allocate _readonlybuffer (which is ignored). So after creating Capacity of such buffer is always 0 (ignoring the actual parameter passed to the constructor) and additional memory is allocated

4) After 3 is fixed, changed code to use DataFrameBuffer constructor with capacity instead of creating empty dataframe buffer and than reallocating memory by calling EnsureCapacity 

___________________________

Result:

Simple tests for 1 million of rows:

```
[GlobalSetup]
public void SetUp()
{
    var values = Enumerable.Range(0, ItemsCount);
    _column1 = new Int32DataFrameColumn("Column1", values);
    _column2 = new Int32DataFrameColumn("Column2", values);
}

[Benchmark]
public void Sum()
{
    var column = _column1 + _column2;
}
```

```
Before PR:
| Method |    Mean |     Error |   StdDev |
|    Sum | 18.02 ms | 0.090 ms | 0.080 ms |

After PR:
| Method |     Mean |     Error |    StdDev |
|    Sum | 6.896 ms | 0.1363 ms | 0.1673 ms |
```

Part of #6824 issue