@asmirnov82 2023-09-14
Running Bencmarks from #6826 

![image](https://github.com/dotnet/machinelearning/assets/83237987/11754bec-3622-42d2-b9ab-6768e691ee31)